Move getThemeFilters
