+++ client/state/selectors/is-ambiguous-theme-filter-term.js
@@ -7,7 +7,7 @@ import { filter, get } from 'lodash';
  * Internal dependencies
  */
 import createSelector from 'lib/create-selector';
-import getThemeFilters from 'state/selectors/get-theme-filters';
+import { getThemeFilters } from 'state/themes/selectors';
 
 import 'state/themes/init';
