+++ client/state/themes/selectors/get-theme-filter-terms-table.js
@@ -7,7 +7,7 @@ import { forIn, keys, mapValues } from 'lodash';
  * Internal dependencies
  */
 import createSelector from 'lib/create-selector';
-import getThemeFilters from 'state/selectors/get-theme-filters';
+import { getThemeFilters } from 'state/themes/selectors/get-theme-filters';
 import isAmbiguousThemeFilterTerm from 'state/selectors/is-ambiguous-theme-filter-term';
 
 import 'state/themes/init';
