+++ client/state/themes/selectors/get-theme-filter-to-term-table.js
@@ -7,7 +7,7 @@ import { forIn, keys, mapValues } from 'lodash';
  * Internal dependencies
  */
 import createSelector from 'lib/create-selector';
-import getThemeFilters from 'state/selectors/get-theme-filters';
+import { getThemeFilters } from 'state/themes/selectors/get-theme-filters';
 import { getThemeFilterTermFromString } from 'state/themes/selectors/get-theme-filter-term-from-string';
 
 import 'state/themes/init';
