+++ client/state/themes/selectors/get-theme-filters.js
@@ -10,6 +10,6 @@ import 'state/themes/init';
  * @param {object}  state Global state tree
  * @returns {object}        A nested list of theme filters, keyed by filter slug
  */
-export default function getThemeFilters( state ) {
+export function getThemeFilters( state ) {
 	return state.themes.themeFilters;
 }
